safari-600.7-branch/LayoutTests/fast/regions/cssom/offsetLeft-offsetTop-inlines-region-in-element.html

<html>
<head>
	<style type="text/css">
		/* Make sure the console and the description don't interfere with the rest of the layout. */
		#description {
			position: absolute;
			top: 0px;
		}

		body {
			border: 10px dotted gray;
			padding: 10px;
		}

		#console {
			position: absolute;
			top: 110px;
		}

		#regionParent {
			border: 1px solid blue;
			border-left: 10px dotted blue;
			border-top: 10px dotted blue;
			width: 600px;
			height: 370px;
			position: relative;
			left: 147px;
			top: 260px;
			margin-left: 25px;
		}

		.region1 {
			-webkit-flow-from: flow1;
			border: 10px dotted red;
			margin: 20px;
			margin-left: 30px;
			width: 404px;
			height: 320px;
		}

		#divMain {
			margin: 5px;
			border: 5px solid blue;
			height: 305px;
			-webkit-flow-into: flow1;
		}

		#divText {
			border: 1px solid green;
			color: green;
			font-size: 14px;
			width: 300px;
			height: 180px;
			position: relative;
			left: 40px;
			top: 12px;
		}

		.grid {
			position: absolute;
			top: 680px;
			width: 100px;
			height: 100px;
			border-right: 1px solid blue;
			text-align: right;
			font-size: 14px;
			color: blue;
		}

		.vgrid {
			position: absolute;
			left: 900px;
			width: 100px;
			height: 100px;
			border-bottom: 1px solid blue;
			text-align: bottom;
			font-size: 14px;
			color: blue;
		}

		#grid1 {
			left: 0px;
		}
		#grid2 {
			left: 100px;
		}
		#grid3 {
			left: 200px;
		}
		#grid4 {
			left: 300px;
		}
		#grid5 {
			left: 400px;
		}
		#grid6 {
			left: 500px;
		}
		#grid7 {
			left: 600px;
		}

		#vgrid1 {
			top: 0px;
		}
		#vgrid2 {
			top: 100px;
		}
		#vgrid3 {
			top: 200px;
		}
		#vgrid4 {
			top: 300px;
		}
		#vgrid5 {
			top: 400px;
		}
		#vgrid6 {
			top: 500px;
		}
		#vgrid7 {
			top: 600px;
		}
		#vgrid8 {
			top: 700px;
			border: none;
		}
	</style>

	<script src="../../../resources/js-test-pre.js"></script>
</head>

<body id="body">
	<div id="divMain">
		<div id="divText"><b>divText</b><br/></div>
		<span class="inlineElements" id="inlineElement" style="color: #000099; position: relative; top: 30px; left: 40px;">Inline element in divMain</span>
	</div>

	<div id="regionParent"><span style="float: right; margin: 10px; color: blue;">This is regionParent</span>
		<div class="region1" id="region"></div>
	</div>

	<div id="grid1" class="grid">100px</div>
	<div id="grid2" class="grid">200px</div>
	<div id="grid3" class="grid">300px</div>
	<div id="grid4" class="grid">400px</div>
	<div id="grid5" class="grid">500px</div>
	<div id="grid6" class="grid">600px</div>
	<div id="grid7" class="grid">700px</div>

	<div id="vgrid1" class="vgrid"></div>
	<div id="vgrid2" class="vgrid">100px</div>
	<div id="vgrid3" class="vgrid">200px</div>
	<div id="vgrid4" class="vgrid">300px</div>
	<div id="vgrid5" class="vgrid">400px</div>
	<div id="vgrid6" class="vgrid">500px</div>
	<div id="vgrid7" class="vgrid">600px</div>
	<div id="vgrid8" class="vgrid">700px</div>

	<script>
		description("Test offsetLeft and offsetTop for blocks and inlines in a named flow when the region's offsetParent is NOT the body.<br/>This test also covers borders.")

		shouldBe("divText.offsetParent", "document.body");
		shouldBe("divText.offsetLeft", "300");
		shouldBe("divText.offsetTop", "350");

		shouldBe("regionParent.offsetLeft", "200");

		shouldBe("region.offsetParent", "regionParent");
		shouldBe("region.offsetLeft", "30");
		shouldBe("region.offsetTop", "20");

		shouldBe("inlineElement.offsetParent", "document.body");
		shouldBe("inlineElement.offsetLeft", "300");
		shouldBe("inlineElement.offsetTop", "550");

		if (window.testRunner) {
			var allGraphicElements = document.querySelectorAll(".region1, #divMain, #divText, .grid, .vgrid, #regionParent, .inlineElements, #regionInfo");
			for (var i=0; i<allGraphicElements.length; i++)
				allGraphicElements[i].style.display = "none";
		}
		
	</script>
	<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
